feat(kitchensink): abstract out editor and actions to be used for media and releases
